OSIS-1757 Add cache for educationgroupyear + remove unusefull parameter + add paramter to cache filter decorator
