+++ base/tests/views/test_education_group.py
@@ -58,222 +58,6 @@ from cms.tests.factories.translated_text import TranslatedTextFactory, Translate
 from cms.tests.factories.translated_text_label import TranslatedTextLabelFactory
 
 
-class EducationGroupSearch(TestCase):
-    @classmethod
-    def setUpTestData(cls):
-        today = datetime.date.today()
-        cls.academic_year = AcademicYearFactory(start_date=today, end_date=today.replace(year=today.year + 1),
-                                                year=today.year)
-        cls.previous_academic_year = AcademicYearFactory(start_date=today.replace(year=today.year - 1),
-                                                         end_date=today - datetime.timedelta(days=1),
-                                                         year=today.year - 1)
-
-        cls.type_training = EducationGroupTypeFactory(category=education_group_categories.TRAINING)
-        cls.type_minitraining = EducationGroupTypeFactory(category=education_group_categories.MINI_TRAINING)
-        cls.type_group = EducationGroupTypeFactory(category=education_group_categories.GROUP)
-
-        oph_entity = EntityFactory()
-        envi_entity = EntityFactory()
-
-        cls.education_group_edph2 = EducationGroupYearFactory(
-            acronym='EDPH2', academic_year=cls.academic_year,
-            partial_acronym='EDPH2_SCS',
-            education_group_type=cls.type_group,
-            management_entity=envi_entity
-        )
-
-        cls.education_group_arke2a = EducationGroupYearFactory(
-            acronym='ARKE2A', academic_year=cls.academic_year,
-            education_group_type=cls.type_training,
-            management_entity=oph_entity
-        )
-
-        cls.education_group_hist2a = EducationGroupYearFactory(
-            acronym='HIST2A', academic_year=cls.academic_year,
-            education_group_type=cls.type_group,
-            management_entity=oph_entity
-        )
-
-        cls.education_group_arke2a_previous_year = EducationGroupYearFactory(
-            acronym='ARKE2A',
-            academic_year=cls.previous_academic_year,
-            education_group_type=cls.type_training,
-            management_entity=oph_entity
-        )
-
-        cls.oph_entity_v = EntityVersionFactory(entity=oph_entity, parent=envi_entity, end_date=None)
-        cls.envi_entity_v = EntityVersionFactory(entity=envi_entity, end_date=None)
-
-        cls.user = PersonFactory().user
-        cls.user.user_permissions.add(Permission.objects.get(codename="can_access_education_group"))
-        cls.url = reverse("education_groups")
-
-    def setUp(self):
-        self.client.force_login(self.user)
-
-    def test_when_not_logged(self):
-        self.client.logout()
-        response = self.client.get(self.url)
-
-        self.assertRedirects(response, "/login/?next={}".format(self.url))
-
-    def test_user_without_permission(self):
-        an_other_user = UserFactory()
-        self.client.force_login(an_other_user)
-        response = self.client.get(self.url)
-
-        self.assertTemplateUsed(response, "access_denied.html")
-        self.assertEqual(response.status_code, HttpResponseForbidden.status_code)
-
-    def test_post_request(self):
-        response = self.client.post(self.url, data={})
-
-        self.assertTemplateUsed(response, "education_group/search.html")
-
-        context = response.context
-        self.assertIsInstance(context["form"], EducationGroupFilter)
-        self.assertEqual(context["object_list_count"], 0)
-        self.assertEqual(context["experimental_phase"], True)
-
-    def test_without_get_data(self):
-        response = self.client.get(self.url)
-
-        self.assertTemplateUsed(response, "education_group/search.html")
-
-        context = response.context
-        self.assertIsInstance(context["form"], EducationGroupFilter)
-        self.assertEqual(context["object_list_count"], 0)
-        self.assertEqual(context["experimental_phase"], True)
-
-    def test_initial_form_data(self):
-        response = self.client.get(self.url)
-
-        form = response.context["form"]
-        self.assertEqual(form.initial["academic_year"], self.academic_year)
-        self.assertEqual(form.initial["category"], education_group_categories.TRAINING)
-
-    def test_with_empty_search_result(self):
-        response = self.client.get(self.url, data={"category": education_group_categories.MINI_TRAINING})
-
-        self.assertTemplateUsed(response, "education_group/search.html")
-
-        context = response.context
-        self.assertIsInstance(context["form"], EducationGroupFilter)
-        self.assertEqual(context["experimental_phase"], True)
-        self.assertEqual(len(context["object_list"]), 0)
-        messages = [str(m) for m in context["messages"]]
-        self.assertIn(_('no_result'), messages)
-
-    def test_search_with_acronym_only(self):
-        response = self.client.get(self.url, data={"acronym": self.education_group_arke2a.acronym})
-
-        self.assertTemplateUsed(response, "education_group/search.html")
-
-        context = response.context
-        self.assertIsInstance(context["form"], EducationGroupFilter)
-        self.assertEqual(context["experimental_phase"], True)
-        self.assertCountEqual(context["object_list"],
-                              [self.education_group_arke2a, self.education_group_arke2a_previous_year])
-
-    def test_search_with_academic_year_only(self):
-        response = self.client.get(self.url, data={"academic_year": self.academic_year.id})
-
-        self.assertTemplateUsed(response, "education_group/search.html")
-
-        context = response.context
-        self.assertIsInstance(context["form"], EducationGroupFilter)
-        self.assertEqual(context["experimental_phase"], True)
-        self.assertCountEqual(context["object_list"],
-                              [self.education_group_arke2a, self.education_group_edph2, self.education_group_hist2a])
-
-    def test_search_with_partial_acronym(self):
-        response = self.client.get(self.url, data={"partial_acronym": self.education_group_edph2.partial_acronym})
-
-        self.assertTemplateUsed(response, "education_group/search.html")
-
-        context = response.context
-        self.assertIsInstance(context["form"], EducationGroupFilter)
-        self.assertEqual(context["experimental_phase"], True)
-        self.assertCountEqual(context["object_list"], [self.education_group_edph2])
-
-    def test_search_with_requirement_entity(self):
-        response = self.client.get(self.url,
-                                   data={"requirement_entity_acronym": self.oph_entity_v.acronym})
-
-        self.assertTemplateUsed(response, "education_group/search.html")
-
-        context = response.context
-        self.assertIsInstance(context["form"], EducationGroupFilter)
-        self.assertEqual(context["experimental_phase"], True)
-        self.assertCountEqual(context["object_list"],
-                              [self.education_group_arke2a, self.education_group_arke2a_previous_year,
-                               self.education_group_hist2a])
-
-    def test_search_with_entities_subordinated(self):
-        response = self.client.get(
-            self.url,
-            data={
-                "requirement_entity_acronym": self.envi_entity_v.acronym,
-                "with_entity_subordinated": True
-            }
-        )
-
-        self.assertTemplateUsed(response, "education_group/search.html")
-
-        context = response.context
-        self.assertIsInstance(context["form"], EducationGroupFilter)
-        self.assertEqual(context["experimental_phase"], True)
-        self.assertCountEqual(
-            context["object_list"],
-            [
-                self.education_group_arke2a,
-                self.education_group_arke2a_previous_year,
-                self.education_group_hist2a,
-                self.education_group_edph2
-            ]
-        )
-
-    def test_search_by_education_group_type(self):
-        response = self.client.get(self.url,
-                                   data={"education_group_type": self.type_group.id})
-
-        self.assertTemplateUsed(response, "education_group/search.html")
-
-        context = response.context
-        self.assertIsInstance(context["form"], EducationGroupFilter)
-        self.assertEqual(context["experimental_phase"], True)
-        self.assertCountEqual(context["object_list"], [self.education_group_hist2a, self.education_group_edph2])
-
-    def test_search_by_education_group_category(self):
-        response = self.client.get(self.url,
-                                   data={"category": education_group_categories.TRAINING})
-
-        self.assertTemplateUsed(response, "education_group/search.html")
-
-        context = response.context
-        self.assertIsInstance(context["form"], EducationGroupFilter)
-        self.assertEqual(context["experimental_phase"], True)
-        self.assertCountEqual(context["object_list"],
-                              [self.education_group_arke2a, self.education_group_arke2a_previous_year])
-
-    def test_with_multiple_criteria(self):
-        response = self.client.get(
-            self.url, data={
-                "academic_year": self.academic_year.id,
-                "acronym": self.education_group_arke2a.acronym,
-                "requirement_entity_acronym": self.envi_entity_v.acronym,
-                "with_entity_subordinated": True
-            }
-        )
-
-        self.assertTemplateUsed(response, "education_group/search.html")
-
-        context = response.context
-        self.assertIsInstance(context["form"], EducationGroupFilter)
-        self.assertEqual(context["experimental_phase"], True)
-        self.assertCountEqual(context["object_list"], [self.education_group_arke2a])
-
-
 class EducationGroupRead(TestCase):
     @classmethod
     def setUpTestData(cls):
